Added error handling closes #60

+++ src/js/controllers/availableForController.js
@@ -4,11 +4,58 @@ const AvailableFor = require('../db/models/AvailableFor');
 
 exports.getAllAvailableFor = async (req, reply) => {
   try {
-    const availableFor = await AvailableFor.findAll();
+    const allAvailableFor = await AvailableFor.findAll();
+
+    if (!allAvailableFor.length) throw 'No items (availableFor) found';
 
     reply.status(200).send({
       status: 'success',
-      results: availableFor.length,
+      results: allAvailableFor.length,
+      data: {
+        allAvailableFor,
+      },
+    });
+  } catch (err) {
+    reply.status(404).send({
+      status: 'fail',
+      message: err,
+    });
+  }
+};
+
+exports.createAvailableFor = async (req, reply) => {
+  try {
+    const newAvailableFor = await AvailableFor.create(req.body);
+
+    if (!newAvailableFor) throw 'Failed to create a new availableFor';
+
+    reply.status(201).send({
+      status: 'success',
+      data: {
+        newAvailableFor,
+      },
+    });
+  } catch (err) {
+    reply.status(400).send({
+      status: 'fail',
+      message: err,
+    });
+  }
+};
+
+exports.getAvailableFor = async (req, reply) => {
+  try {
+    const availableFor = await AvailableFor.findOne({
+      where: {
+        id: req.params.id,
+      },
+    });
+
+    if (!availableFor)
+      throw 'availableFor with the specified ID does not exists';
+
+    reply.status(200).send({
+      status: 'success',
       data: {
         availableFor,
       },
@@ -21,51 +68,13 @@ exports.getAllAvailableFor = async (req, reply) => {
   }
 };
 
-exports.createAvailableFor = async (req, res) => {
+exports.updateAvailableFor = async (req, reply) => {
   try {
-    const newAvailableFor = await AvailableFor.create(req.body);
-
-    res.status(201).send({
-      status: 'success',
-      data: {
-        newAvailableFor,
-      },
-    });
-  } catch (err) {
-    res.status(400).send({
-      status: 'fail',
-      message: err,
-    });
-  }
-};
-
-exports.getAvailableFor = async (req, res) => {
-  try {
-    const availableFor = await AvailableFor.findOne({
-      where: {
-        id: req.params.id,
-      },
-    });
-
-    res.status(200).send({
-      status: 'success',
-      data: {
-        availableFor,
-      },
-    });
-  } catch (err) {
-    res.status(404).send({
-      status: 'fail',
-      message: err,
-    });
-  }
-};
-
-exports.updateAvailableFor = async (req, res) => {
-  try {
-    await AvailableFor.update(req.body, {
+    const [updated] = await AvailableFor.update(req.body, {
       where: { id: req.params.id },
     });
+
+    if (!updated) throw 'availableFor with the specified ID does not exists';
 
     const updatedAvailableFor = await AvailableFor.findOne({
       where: {
@@ -73,34 +82,41 @@ exports.updateAvailableFor = async (req, res) => {
       },
     });
 
-    res.status(200).send({
+    reply.status(200).send({
       status: 'success',
       data: {
         updatedAvailableFor,
       },
     });
   } catch (err) {
-    res.status(404).send({
+    reply.status(404).send({
       status: 'fail',
       message: err,
     });
   }
 };
 
-exports.deleteAvailableFor = async (req, res) => {
+exports.deleteAvailableFor = async (req, reply) => {
   try {
+    const availableFor = await AvailableFor.findOne({
+      where: { id: req.params.id },
+    });
+
+    if (!availableFor)
+      throw 'availableFor with the specified ID does not exists';
+
     await AvailableFor.destroy({
       where: { id: req.params.id },
     });
 
-    res.status(200).send({
+    reply.status(200).send({
       status: 'success',
       data: null,
     });
   } catch (err) {
-    res.status(404).send({
+    reply.status(404).send({
       status: 'fail',
-      message: err,
+      message: err.message,
     });
   }
 };
